+++ find_a_supplier/forms.py
@@ -5,7 +5,7 @@ from django.utils.safestring import mark_safe
 from captcha.fields import ReCaptchaField
 from directory_constants import choices, urls
 from directory_components import forms
-from directory_forms_api_client.forms import GovNotifyActionMixin
+from directory_forms_api_client.forms import GovNotifyEmailActionMixin
 from directory_validators.common import not_contains_url_or_email
 from django.core.validators import EMPTY_VALUES
 
@@ -82,7 +82,7 @@ class CompanySearchForm(forms.Form):
         return self.cleaned_data['page'] or self.fields['page'].initial
 
 
-class ContactCompanyForm(GovNotifyActionMixin, forms.Form):
+class ContactCompanyForm(GovNotifyEmailActionMixin, forms.Form):
     TERMS_CONDITIONS_MESSAGE = (
         'Tick the box to confirm you agree to the terms and conditions.'
     )
@@ -121,7 +121,7 @@ class ContactCompanyForm(GovNotifyActionMixin, forms.Form):
         validators=[not_contains_url_or_email],
     )
     email_address = forms.EmailField(
-        label='Your email address:',
+        label='Your email address',
     )
     sector = forms.ChoiceField(
         label='Your industry',
@@ -212,7 +212,7 @@ def serialize_anonymous_subscriber_forms(cleaned_data):
     }
 
 
-class ContactForm(GovNotifyActionMixin, forms.Form):
+class ContactForm(GovNotifyEmailActionMixin, forms.Form):
 
     def __init__(self, industry_choices, *args, **kwargs):
         super().__init__(*args, **kwargs)